requirements.txt
faster-whisper==1.0.3
ffmpeg-python==0.2.0